Extract interfaces and classes to separete files.
