+++ csharp-advanced-topics/ActionAndFuncInCsharp/ActionAndFuncInCsharp/Implementations/ParkingLot.cs
@@ -1,0 +1,12 @@
+﻿using ActionAndFuncInCsharp.Interfaces;
+
+namespace ActionAndFuncInCsharp.Implementations
+{
+    public class ParkingLot : IParkingLot
+    {
+        public string Park(IVehicle vehicle)
+        {
+            return vehicle.Park();
+        }
+    }
+}
