+++ csharp-advanced-topics/ActionAndFuncInCsharp/ActionAndFuncInCsharp/Implementations/Progress.cs
@@ -1,0 +1,18 @@
+﻿using ActionAndFuncInCsharp.Interfaces;
+
+namespace ActionAndFuncInCsharp.Implementations
+{
+    public class Progress : IProgress
+    {
+        public string Message { get; }
+        public double Value { get; }
+        public string BarDrawing { get; }
+
+        public Progress(string message, double value, string barDrawing)
+        {
+            Message = message;
+            Value = value;
+            BarDrawing = barDrawing;
+        }
+    }
+}
