+++ csharp-advanced-topics/ActionAndFuncInCsharp/ActionAndFuncInCsharp/Implementations/Vehicles/Ambulance.cs
@@ -1,0 +1,12 @@
+﻿using ActionAndFuncInCsharp.Interfaces;
+
+namespace ActionAndFuncInCsharp.Implementations.Vehicles
+{
+    public class Ambulance : IVehicle
+    {
+        public string Park()
+        {
+            return "The Ambulance is parked.";
+        }
+    }
+}
