+++ csharp-advanced-topics/ActionAndFuncInCsharp/ActionAndFuncInCsharp/Implementations/Vehicles/Bus.cs
@@ -1,0 +1,12 @@
+﻿using ActionAndFuncInCsharp.Interfaces;
+
+namespace ActionAndFuncInCsharp.Implementations.Vehicles
+{
+    public class Bus : IVehicle
+    {
+        public string Park()
+        {
+            return "The Bus is parked.";
+        }
+    }
+}
