+++ csharp-advanced-topics/ActionAndFuncInCsharp/ActionAndFuncInCsharp/Implementations/Vehicles/Car.cs
@@ -1,0 +1,12 @@
+﻿using ActionAndFuncInCsharp.Interfaces;
+
+namespace ActionAndFuncInCsharp.Implementations.Vehicles
+{
+    public class Car : IVehicle
+    {
+        public string Park()
+        {
+            return "The Car is parked.";
+        }
+    }
+}
