+++ csharp-advanced-topics/ActionAndFuncInCsharp/ActionAndFuncInCsharp/Implementations/Vehicles/Motorcycle.cs
@@ -1,0 +1,12 @@
+﻿using ActionAndFuncInCsharp.Interfaces;
+
+namespace ActionAndFuncInCsharp.Implementations.Vehicles
+{
+    public class Motorcycle : IVehicle
+    {
+        public string Park()
+        {
+            return "The Motorcycle is parked.";
+        }
+    }
+}
