+++ csharp-advanced-topics/ActionAndFuncInCsharp/ActionAndFuncInCsharp/Implementations/Vehicles/Truck.cs
@@ -1,0 +1,12 @@
+﻿using ActionAndFuncInCsharp.Interfaces;
+
+namespace ActionAndFuncInCsharp.Implementations.Vehicles
+{
+    public class Truck : IVehicle
+    {
+        public string Park()
+        {
+            return "The Truck is parked.";
+        }
+    }
+}
