+++ csharp-advanced-topics/ActionAndFuncInCsharp/ActionAndFuncInCsharp/Interfaces/IParkingLot.cs
@@ -1,0 +1,7 @@
+﻿namespace ActionAndFuncInCsharp.Interfaces
+{
+    public interface IParkingLot
+    {
+        public string Park(IVehicle vehicle);
+    }
+}
