+++ csharp-advanced-topics/ActionAndFuncInCsharp/ActionAndFuncInCsharp/Interfaces/IProgress.cs
@@ -1,0 +1,9 @@
+﻿namespace ActionAndFuncInCsharp.Interfaces
+{
+    public interface IProgress
+    {
+        public string Message { get; }
+        public string BarDrawing { get; }
+        public double Value { get; }
+    }
+}
