+++ csharp-advanced-topics/ActionAndFuncInCsharp/ActionAndFuncInCsharp/Interfaces/IProgressManager.cs
@@ -1,0 +1,11 @@
+﻿namespace ActionAndFuncInCsharp.Interfaces
+{
+    public interface IProgressManager
+    {
+        public string CurrentMessage { get; }
+        public string CurrentBarDrawing { get; }
+        public string CurrentProgressValue { get; }
+        public string GetProgressBarDrawing(double value);
+        public void UpdateProgress(IProgress progress);
+    }
+}
