+++ csharp-advanced-topics/ActionAndFuncInCsharp/ActionAndFuncInCsharp/Interfaces/IVehicle.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ActionAndFuncInCsharp.Interfaces
+{
+    public interface IVehicle
+    {
+        public string Park();
+    }
+}
