+++ csharp-advanced-topics/ActionAndFuncInCsharp/Tests/ParkingManagementUnitTests.cs
@@ -1,3 +1,6 @@
+using ActionAndFuncInCsharp.Implementations;
+using ActionAndFuncInCsharp.Implementations.Vehicles;
+using ActionAndFuncInCsharp.Interfaces;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
 
@@ -59,8 +62,9 @@ namespace Tests
         {
             //Arrange                        
             const string isParkedMessage = "The Bus is parked.";
-            var mockedBus = new Mock<IVehicle>();
+            var mockedBus = new Mock<IVehicle>();            
             mockedBus.Setup(x => x.Park()).Returns(isParkedMessage);
+            
             var mockedVehicles = new List<IVehicle>() { mockedBus.Object };
             var parkingLot = new ParkingLot();
 
@@ -79,17 +83,19 @@ namespace Tests
             //Arrange                                    
             var mockedTruck = new Mock<IVehicle>();
             mockedTruck.Setup(x => x.Park());
+            
             var mockedVehicles = new List<IVehicle>() { mockedTruck.Object };            
             var mockedParkingLot = new Mock<IParkingLot>();
             mockedParkingLot.Setup(x => x.Park(It.IsAny<IVehicle>()));                
+            
             var mockedProgressManager = new Mock<IProgressManager>();
             mockedProgressManager.Setup(x => x.GetProgressBarDrawing(It.IsAny<double>()));
             mockedProgressManager.Setup(x => x.UpdateProgress(It.IsAny<IProgress>()));
             mockedProgressManager.Setup(x => x.CurrentMessage);
             mockedProgressManager.Setup(x => x.CurrentBarDrawing);
             mockedProgressManager.Setup(x => x.CurrentProgressValue);
+            
             var parkingManager = new ParkingManager();
-
 
             //Act
             Exception? actualException = null;
